+++ secundarios.h
@@ -1,0 +1,38 @@
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <ctype.h>
+#define DIMS 80
+#define MAX 10
+
+typedef struct nodostring{
+    char cadena[DIMS];
+    struct nodostring *sig;
+}nodostring;
+typedef nodostring *tlistastring;
+
+typedef char nombre[MAX];
+
+typedef struct{
+    char rotulo[DIMS];
+    char mnemonico[MAX];
+    nombre argumentos[4];
+    char comentario[DIMS];
+}registroinstruccion;
+
+typedef struct nodo{
+    char rotulo[DIMS];
+    int numerodelinea;
+    struct nodo *sig;
+}nodo;
+typedef nodo* tlista;
+
+//proce y func auxiliares
+int identificaBase(char);
+void agregarotulo(tlista*,char[],int);
+int codigooperando(char[]);
+int basebtodecimal(char[],int);
+int potencia(int,int);
+void mayus(char[]);
+void muestramemoria(int[]);//despues borrar
+void muestralista(tlistastring);
